# Conflicts:
#	.idea/.gitignore
#	.idea/modules.xml
#	.idea/vcs.xml

+++ src/main/java/com/ensak/petines/repositories/CommandeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface CommandeRepository extends CrudRepository<Commande, Long> {
+public interface CommandeRepository extends CrudRepository<Commande, Integer> {
 
     List<Commande> findAll();
